requirements.txt
asyncio
bs4
datetime
discord.py
python_dotenv
pyyaml
requests
virtualenv
pandas